+++ uui-components/src/inputs/NumericInput.tsx
@@ -203,7 +203,6 @@ export const NumericInput = React.forwardRef<HTMLDivElement, NumericInputProps>(
 
     return (
         <div
-            data-testid="uui-numeric-input-wrapper"
             className={ cx(
                 css.container,
                 uuiElement.inputBox,
@@ -223,7 +222,6 @@ export const NumericInput = React.forwardRef<HTMLDivElement, NumericInputProps>(
             { ...props.rawProps }
         >
             <input
-                data-testid="uui-numeric-input"
                 type="number"
                 className={ cx(uuiElement.input, props.inputCx, props.align === 'right' && css.alignRight, isPlaceholderColored && uuiElement.valueInPlaceholder) }
                 disabled={ props.isDisabled }
